Networkservices: Further cleanup and consolidation, get rid of the almost useless IntNetIf helper class and add methods to the low level code offering the missing functionality there, bugref:10297 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@97079 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/NetworkServices/NAT/VBoxNetLwipNAT.cpp
@@ -1953,7 +1953,7 @@ VBoxNetLwipNAT::netifLinkoutput(netif *pNetif, pbuf *pPBuf) RT_NOTHROW_DEF
 
     size_t cbFrame = (size_t)pPBuf->tot_len - ETH_PAD_SIZE;
     INTNETFRAME Frame;
-    rc = IntNetR3IfQueryOutputFrame(self->m_hIf, cbFrame, &Frame);
+    rc = IntNetR3IfQueryOutputFrame(self->m_hIf, (uint32_t)cbFrame, &Frame);
     if (RT_FAILURE(rc))
         return ERR_MEM;
